refactor: rename and reorganize equilibrium state initialization components
