+++ src/laser_measles/components/base_initialize_equilibrium_states.py
@@ -5,28 +5,28 @@ Component for initializing the population in each of the model states by rough e
 from pydantic import BaseModel
 from pydantic import Field
 
-from laser_measles.base import BaseComponent
+from laser_measles.base import BasePhase
 from laser_measles.base import BaseLaserModel
 
 
-class InitializeEquilibriumStatesParams(BaseModel):
+class BaseInitializeEquilibriumStatesParams(BaseModel):
     """
     Parameters for the InitializeEquilibriumStatesProcess.
     """
 
-    R0: float = Field(default=8.0, description="Basic reproduction number", ge=0.0)
+    R0: float = Field(default=8.0, description="Basic reproduction number setting the initialization", ge=0.0)
 
 
-class InitializeEquilibriumStatesProcess(BaseComponent):
+class BaseInitializeEquilibriumStatesProcess(BasePhase):
     """
     Initialize S, R states of the population in each of the model states by rough equilibrium of R0.
     """
 
-    def __init__(self, model: BaseLaserModel, verbose: bool = False, params: InitializeEquilibriumStatesParams | None = None):
+    def __init__(self, model: BaseLaserModel, verbose: bool = False, params: BaseInitializeEquilibriumStatesParams | None = None):
         super().__init__(model, verbose)
-        self.params = params or InitializeEquilibriumStatesParams()
+        self.params = params or BaseInitializeEquilibriumStatesParams()
 
-    def initialize(self, model: BaseLaserModel):
+    def _initialize(self, model: BaseLaserModel):
         """
         Initialize the population in each of the model states by rough equilibrium of R0.
         THis is run after model.run()
